workstation/qbittorrent: remove
Some gcc breaking change broke the program I was relying on, and I
cannot be bothered with patching it. Instead I'll just remove the
module, since it was pretty bad, and I don't really download anything on
windows anymore. The drive is still shared, so if I do want to, I can
just do it on linux, and reboot after.