treewide: Hyprland -> sway
I've wanted to move off Hyprland for a while now. I have multiple issues
with it (some of them very annoying), that aren't likely to be fixed.
There are also just things that work better in Sway. I value simplicity
and correctness, both of which Sway is better at (they might overdo it,
but that's still better than the Hyprland situation).