treewide: remove SVP remnants
I haven't used SVP in a while, and the bits in the mpv config turned
out to break curd (because it uses the ipc to communicate with mpv,
and the path in my config took precedence), which I do use. Apparently
SVP also introduced some changes, so that the socket path doesn't have
to be dynamic (therefore allowing for multiple mpv instances). I'll
just figure this out if I want to use SVP again.