treewide: devimify
I've been using Emacs for a while now, and I'm ready to commit to it
completely. I've come to realize that I'm not really a fan of global
editing. Switching between modes feels like it adds a lot of
overhead. This isn't a real argument, since I've already learnt most
of the Vim keybinds, but Emacs lets you do the same things with less
keybinds overall (e.g., instead of motions, you just use a mark or a
region).

This commit removes the following:
- LazyVim
- nvf
- Neovide
- home/less -> workstation/less