workstation: emacs -> helix
Emacs is nice, but minor issues kept compounding, so it's time to try something new. The tight, frictionless integration and bloatless feel makes up for the lack of plugin support,  and I honestly haven't been missing that many features. For note taking, I will be using Typst, and for merge resolution (support is planned for Helix), I will try Meld.